# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Voice AI Sales Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiKeyInput {
  /**
     * ElevenLabs API key
     * @minLength 1
     */
  apiKey: string;
}

export interface SaveKeyResult {
  success: boolean;
  message: string;
}

export interface SettingsStatus {
  keySaved: boolean;
  /**
     * Partially masked key for display (e.g. "sk_...xxxx"), never the full key
     * @nullable
     */
  maskedKey?: string | null;
}

export interface AgentCreateResult {
  success: boolean;
  agentId: string;
  message?: string;
}

/**
 * Sales direction used to configure the agent
 */
export type AgentCreateInputSalesMode = typeof AgentCreateInputSalesMode[keyof typeof AgentCreateInputSalesMode];


export const AgentCreateInputSalesMode = {
  online_course: 'online_course',
  fitness_membership: 'fitness_membership',
  crm_system: 'crm_system',
} as const;

export interface AgentCreateInput {
  /** Sales direction used to configure the agent */
  salesMode: AgentCreateInputSalesMode;
  /**
     * Optional text extracted from the uploaded service brief
     * @maxLength 50000
     */
  serviceDescription?: string;
  /**
     * Original name of the uploaded service brief
     * @maxLength 255
     * @nullable
     */
  fileName?: string | null;
}

/**
 * Sales direction configured for the current agent
 * @nullable
 */
export type AgentStatusSalesMode = typeof AgentStatusSalesMode[keyof typeof AgentStatusSalesMode] | null;


export const AgentStatusSalesMode = {
  online_course: 'online_course',
  fitness_membership: 'fitness_membership',
  crm_system: 'crm_system',
} as const;

export interface AgentStatus {
  hasAgent: boolean;
  /** @nullable */
  agentId?: string | null;
  /**
     * Sales direction configured for the current agent
     * @nullable
     */
  salesMode?: AgentStatusSalesMode;
  /**
     * Name of the service description file attached to the current agent
     * @nullable
     */
  fileName?: string | null;
  /**
     * Text of the service description attached to the current agent
     * @nullable
     */
  serviceDescription?: string | null;
}

export type AgentConfigurationInputSalesMode = typeof AgentConfigurationInputSalesMode[keyof typeof AgentConfigurationInputSalesMode];


export const AgentConfigurationInputSalesMode = {
  online_course: 'online_course',
  fitness_membership: 'fitness_membership',
  crm_system: 'crm_system',
} as const;

export interface AgentConfigurationInput {
  salesMode: AgentConfigurationInputSalesMode;
  /**
     * @maxLength 255
     * @nullable
     */
  fileName?: string | null;
  /**
     * @maxLength 50000
     * @nullable
     */
  serviceDescription?: string | null;
}

export interface SignedUrlResult {
  /** Signed WebSocket URL for ElevenLabs conversation */
  signedUrl: string;
}

export type ClientEventInputLevel = typeof ClientEventInputLevel[keyof typeof ClientEventInputLevel];


export const ClientEventInputLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
} as const;

export interface ClientEventInput {
  level: ClientEventInputLevel;
  message: string;
  /**
     * Optional JSON-serialized data payload
     * @nullable
     */
  data?: string | null;
}

export interface LogResult {
  ok: boolean;
}

export interface ErrorResult {
  error: string;
}