# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Voice AI Sales Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/create:
    post:
      operationId: createAgent
      tags: [agent]
      summary: Create ElevenLabs AI sales agent
      description: Creates a new AI sales agent via ElevenLabs API using the saved key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreateInput"
      responses:
        "200":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentCreateResult"
        "400":
          description: Error creating agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
...
  /agent/configuration:
    patch:
      operationId: updateAgentConfiguration
      tags: [agent]
      summary: Update the current agent configuration
      description: Updates the saved ElevenLabs agent prompt and its local configuration metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentConfigurationInput"
      responses:
        "200":
          description: Agent configuration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
        "400":
          description: Error updating agent configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
...
components:
  schemas:
...
    AgentCreateResult:
      type: object
      required: [success, agentId]
      properties:
        success:
          type: boolean
        agentId:
          type: string
        message:
          type: string
    AgentCreateInput:
      type: object
      required: [salesMode]
      properties:
        salesMode:
          type: string
          enum: [online_course, fitness_membership, crm_system]
          description: Sales direction used to configure the agent
        serviceDescription:
          type: string
          maxLength: 50000
          description: Optional text extracted from the uploaded service brief
        fileName:
          type: ["string", "null"]
          maxLength: 255
          description: Original name of the uploaded service brief
    AgentStatus:
      type: object
      required: [hasAgent]
      properties:
        hasAgent:
          type: boolean
        agentId:
          type: ["string", "null"]

        salesMode:
          type: ["string", "null"]
          enum: [online_course, fitness_membership, crm_system, null]
          description: Sales direction configured for the current agent

        fileName:
          type: ["string", "null"]
          description: Name of the service description file attached to the current agent

        serviceDescription:
          type: ["string", "null"]
          description: Text of the service description attached to the current agent
    AgentConfigurationInput:
      type: object
      required: [salesMode]
      properties:
        salesMode:
          type: string
          enum: [online_course, fitness_membership, crm_system]
        fileName:
          type: ["string", "null"]
          maxLength: 255
        serviceDescription:
          type: ["string", "null"]
          maxLength: 50000
...
    ErrorResult:
      type: object
      required: [error]
      properties:
        error:
          type: string