# lib/api-zod/src/generated/types/agentConfigurationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Voice AI Sales Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentConfigurationInputSalesMode } from './agentConfigurationInputSalesMode';

export interface AgentConfigurationInput {
  salesMode: AgentConfigurationInputSalesMode;
  /**
     * @maxLength 255
     * @nullable
     */
  fileName?: string | null;
  /**
     * @maxLength 50000
     * @nullable
     */
  serviceDescription?: string | null;
}