# lib/api-zod/src/generated/types/agentCreateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Voice AI Sales Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentCreateInputSalesMode } from './agentCreateInputSalesMode';

export interface AgentCreateInput {
  /** Sales direction used to configure the agent */
  salesMode: AgentCreateInputSalesMode;
  /**
     * Optional text extracted from the uploaded service brief
     * @maxLength 50000
     */
  serviceDescription?: string;
  /**
     * Original name of the uploaded service brief
     * @maxLength 255
     * @nullable
     */
  fileName?: string | null;
}